requirements.txt
dj-database-url==0.5.0
Django==2.1.2
mysqlclient==1.3.13
pytz==2018.6
django-versatileimagefield==1.10
